Offer accessor for optflags without -flto
Factory is in progress of enabling -flto by default. Not all code supports this, so
those need an easy way to opt out from -flto. As such, we offer a new macro stripping
-flto from the default flags.

It can be used in a spec file the same way optflags is often used, e.g.

export CFLAGS="%{optflagsnolto}"