Add %fillup_requires, %install_info_requires, %permissions_requires
The current %fillup_prereq and %install_info_prereq are only packages
names, and they are often used together with a PreReq line in plenty
of Tumbleweed .spec files. Dependencies on the "permissions" package
is completely open-coded at this time.

Not in all instances is a hard PreReq required, and centralizing that
in a macro that expands to not just package names, but also the
dependency tag name, seems worthwhile. This is what the three new
macros hereby introduced are for.