Add macros for locale provides
For packages that ship mo files, add automatic provides for the
corresponding locale instead of an unconditional Supplements.
That way the lang subpackage won't be installed on English only
installations.
The mechanism requires to stick to the common openSUSE naming
convention that a package "foo" has it's translations in a package
called "foo-lang".

Requires https://github.com/rpm-software-management/rpm/pull/575