Factor-out userspace live patching macros
Userspace live patching requires that packages are built with special
flags and artifacts to be packed for later usee. Instead of spawning
code for this in every .spec file, a much better way is to factor out
everything in a common file.

The way it is intended to be used is the following: packages that
userspace livepatching support is desired should have defined in
its begining:

%define livepatchable 1

Then it should append %pack_ipa_dumps on %install.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>
Co-authored-by: Martin Liska <mliska@suse.de>